Split code into many components
