+++ src/App.css
@@ -33,7 +33,9 @@ body{
 
 .RichEditor-blockquote {
   color: #999;
-  font-family: 'Hoefler Text', Georgia, serif;
+  /*font-family: 'Hoefler Text', Georgia, serif;*/
+  font-family: "Comic Sans MS", cursive, sans-serif;
+  font-size: 1em;
   font-style: italic;
   text-align: center;
 }
